moved models to research_hub directory
