+++ network-api/networkapi/wagtailpages/tests/research_hub/test_detail_page.py
@@ -11,7 +11,7 @@ from networkapi.wagtailpages.models import (
     PublicationPage,
     ResearchDetailPage,
 )
-from networkapi.wagtailpages.pagemodels.research_hub import authors_index
+from networkapi.wagtailpages.pagemodels.libraries.research_hub import authors_index
 from networkapi.wagtailpages.tests.research_hub import base as research_test_base
 from networkapi.wagtailpages.tests.research_hub import utils as research_test_utils
